First pass to use python setuptools to unpack kalite source
Previously, we created an archive of the kalite source (and its dependencies)
but because I didn't know anything about setuptools I wasn't using it properly.

Now instead of copying the whole source, we reduce the responsibility of
the windows installer by copying the source archive and unpacking it using
git. This means that the source now lives in the python site-packages instead
of the user-chosen installation directory. That directory now just has the
source archive and systray executable files.

We no longer set KALITE_ROOT_DATA_PATH, since it will now be unpacked
in the expected location by pip (namely sys.prefix + "share\kalite").

Since we don't set any environment variables, don't force a restart either.

The systray executable will need to be made aware of the location of the
`kalite.bat` script, since it no longer lives in the application directory.

We may discover more tasks as development continues.