Make systray program get kalite.bat path from env
Since the `kalite.bat` file is *not* installed by pip to the KA Lite application
directory, the systray program won't find it there. Instead, we'll have to set the
path of that script at install time using an environment variable -- I've chosen
KALITE_SCRIPT_DIR.

The program attempts to read this directory when calling `kalite.bat`. On error
it sends a message to the systray program.