Expand KALITE_SCRIPT_DIR properly when creating startup task
Don't use the external script, which relies on the expansion of
KALITE_SCRIPT_DIR -- the env won't be updated until *after* the
installer exits, so the scheduled task will be misconfigured.

In the installer, create the task from the cmd line and fetch
KALITE_SCRIPT_DIR directly from the registry.