Enabling system start task prompts user for password
This is because Windows will refuse to start the KA Lite server
in the background *before* the user logs in if the password is
not provided. Note that passing the /np (no password) flag will
cause the task to *attempt* to start, but it will fail because
"no logon server could process the request".

If this script is called from the tasktray program, a command
prompt will appear prompting the user for the password.