Get rid of getversion.bat
It depends on the ka-lite repo being in the installers repo.
But that's not the case if we can grab the sdist directly from PyPI.
Instead get the version from an environment variable, and update
the readme accordingly.

The change from ShellExec to Exec was a bugfix -- ShellExec returns
"True" even if the program fails, and only returns "False" if the
program *fails to open*. So it wasn't catching a critical failure.